auth-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
aiosqlite>=0.21.0
alembic>=1.14.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.20
httpx>=0.28.0
eval_type_backport>=0.2.0
